003-stack_cloud_functions-main/cf_vision_api/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.2.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-cloud-vision==3.4.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-http-client==3.3.7
requests==2.28.2
rsa==4.9
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
urllib3==1.26.14
watchdog==2.3.1
Werkzeug==2.2.3
zipp==3.15.0
